1. добавлены репозитории и сервисные слои для работы с ними 2. добавлена возможность бронирования вещи 3. добавлена обработка ошибок 4. добавлена возможность оставлять отзывы по прошедшим бронированиям 5. произведен рефакторинг всего кода

+++ src/main/java/ru/practicum/shareit/item/storage/ItemStorage.java
@@ -15,17 +15,17 @@ public interface ItemStorage {
     /**
      * проверка наличия вещи с @param id в хранилище
      */
-    boolean exists(int id);
+    boolean exists(long id);
 
     /**
      * получение всех вещей из хранилища
      */
-    List<Item> getAll(int userId);
+    List<Item> getAll(long userId);
 
     /**
      * получение определенной вещи с @param id из хранилища
      */
-    Item get(int id);
+    Item get(long id);
 
     /**
      * обновление данных определенной вещи в хранилище
@@ -40,5 +40,5 @@ public interface ItemStorage {
     /**
      * удаление вещи с @param id из хранилища
      */
-    void delete(int id);
+    void delete(long id);
 }
